backend: modificare date trimise pentru grafic frontend: modificare date primite pentru grafic

+++ application/src/main/java/health/tracking/application/dto/DashboardWeekDTO.java
@@ -14,5 +14,5 @@ import java.util.List;
 public class DashboardWeekDTO {
     List<Integer> calories;
     List<Integer>qualitySleep;
-    List<Integer> caloriesConsumed;
+    List<Double> caloriesConsumed;
 }
